Mod: flea-auth授权POJO工具类拆分优化
