+++ .gitignore
@@ -23,3 +23,4 @@ hs_err_pid*
 .idea/
 *.iml
 target/
+/.workbuddy
